feat(接口测试): 场景设置的步骤等待时间

+++ backend/services/project-management/src/main/java/io/metersphere/project/api/processor/TimeWaitingProcessor.java
@@ -15,5 +15,5 @@ public class TimeWaitingProcessor extends MsProcessor {
      * 等待时间
      * 单位：毫秒
      */
-    private Integer delay;
+    private Long delay;
 }
